+++ packages/api/lib/controllers/measurementsController.js
@@ -19,6 +19,7 @@ const
  * @apiGroup Measurements
  * @apiName getLatestMeasurements
  * @apiUse BoxIdParam
+ * @apiParam {NumberNumber=1-100} [count] Number of measurements to be retrieved for every sensor.
  */
 /**
  * @api {get} /boxes/:senseBoxId/sensors/:sensorId Get latest measurements of a sensor
@@ -35,7 +36,30 @@ const getLatestMeasurements = async function getLatestMeasurements (req, res, ne
   let box;
 
   try {
-    box = await Box.findBoxById(req._userParams.boxId, { onlyLastMeasurements: true });
+    if (req._userParams.count) {
+      box = await Box.findBoxById(req._userParams.boxId, {
+        populate: false,
+        onlyLastMeasurements: false,
+        count: req._userParams.count,
+        projection: {
+          name: 1,
+          lastMeasurementAt: 1,
+          sensors: 1,
+          grouptag: 1
+        }
+      });
+
+      const measurements = await Measurement.findLatestMeasurementsForSensorsWithCount(box, req._userParams.count);
+      for (let index = 0; index < box.sensors.length; index++) {
+        const sensor = box.sensors[index];
+        const values = measurements.find(elem => elem._id.equals(sensor._id));
+        sensor['lastMeasurements'] = values;
+      }
+    } else {
+      box = await Box.findBoxById(req._userParams.boxId, {
+        onlyLastMeasurements: true
+      });
+    }
   } catch (err) {
     handleError(err, next);
 
@@ -200,6 +224,46 @@ const getDataMulti = async function getDataMulti (req, res, next) {
 
     if (download === 'true') {
       res.header('Content-Disposition', `attachment; filename=${createDownloadFilename(req.date(), 'download', [phenomenon, ...columns], format)}`);
+    }
+
+    stream
+      .pipe(res);
+  } catch (err) {
+    handleError(err, next);
+  }
+};
+
+
+/**
+ * @api {get,post} /boxes/data?grouptag=:grouptag Get latest measurements for a grouptag as JSON
+ * @apiDescription Download data of a given grouptag from multiple senseBoxes as JSON
+ * @apiGroup Measurements
+ * @apiName getDataByGroupTag
+ * @apiParam {String} grouptag The grouptag to search by.
+ */
+const getDataByGroupTag = async function getDataByGroupTag (req, res, next) {
+  const { grouptag, format } = req._userParams;
+  const queryTags = grouptag.split(',');
+  // build query
+  const queryParams = {};
+  if (grouptag) {
+    queryParams['grouptag'] = { '$all': queryTags };
+  }
+
+  try {
+    let stream = await Box.findMeasurementsOfBoxesByTagStream({
+      query: queryParams
+    });
+    stream = stream
+      .on('error', function (err) {
+        return handleError(err, next);
+      });
+    switch (format) {
+    case 'json':
+      res.header('Content-Type', 'application/json');
+      stream = stream
+        .pipe(jsonstringify({ open: '[', close: ']' }));
+      break;
     }
 
     stream
@@ -409,10 +473,18 @@ module.exports = {
     validateFromToTimeParams,
     getDataMulti
   ],
+  getDataByGroupTag: [
+    retrieveParameters([
+      { name: 'grouptag', required: true },
+      { name: 'format', defaultValue: 'json', allowedValues: ['json'] }
+    ]),
+    getDataByGroupTag
+  ],
   getLatestMeasurements: [
     retrieveParameters([
       { predef: 'boxId', required: true },
       { predef: 'sensorId' },
+      { name: 'count', dataType: 'Integer', min: 1, max: 100, required: false },
       { name: 'onlyValue', required: false }
     ]),
     getLatestMeasurements