+++ packages/api/lib/helpers/userParamHelpers.js
@@ -348,12 +348,36 @@ const retrieveLocationParameter = function ({ value }) {
   }
 };
 
+const retrieveNearParameter = function ({ value, dataType, dataTypeIsArray }) {
+  try {
+    // wrap dataType in array for calling of casting function
+    dataType = dataTypeIsArray ? dataType[0] : dataType;
+    if (!dataTypeIsArray && Array.isArray(value)) {
+      return { error: ARRAY_NOT_ALLOWED };
+    }
+
+    // test and cast value against dataType
+    value = castParam(value, dataType, dataTypeIsArray);
+
+    if (typeof value === 'undefined') {
+      return { error: CAST_FAILED };
+    }
+
+    return { castedValue: transformAndValidateCoords(value) };
+  } catch (err) {
+    return { error: ERROR_CUSTOM_MESSAGE, message: err.message };
+  }
+};
+
 const
   GET_DATA_MULTI_DEFAULT_COLUMNS = ['sensorId', 'createdAt', 'value', 'lat', 'lon'],
   GET_DATA_MULTI_ALLOWED_COLUMNS = ['createdAt', 'value', 'lat', 'lon', 'height', 'unit', 'boxId', 'sensorId', 'phenomenon', 'sensorType', 'boxName', 'exposure'];
 
 // functions which receive req._userparams as parameter
 const retrieveParametersPredefs = {
+  'dateNoDefault' () {
+    return { name: 'date', dataType: 'RFC 3339' };
+  },
   'fromDate' ({ toDate }) {
     if (!toDate) {
       return {};
@@ -386,7 +410,7 @@ const retrieveParametersPredefs = {
     return { name: 'bbox', dataType: 'bbox' };
   },
   'near' () {
-    return { name: 'near', dataType: 'as-is' };
+    return { name: 'near', dataType: ['Number'], paramValidatorAndParser: retrieveNearParameter };
   },
   'maxDistance' () {
     return { name: 'maxDistance', dataType: 'as-is' };
@@ -512,6 +536,18 @@ const parseAndValidateTimeParamsForFindAllBoxes = function parseAndValidateTimeP
   next();
 };
 
+const validateDateNotPast = function validateDateNotPast (req, res, next) {
+  if (req._userParams.date) {
+    const { date } = req._userParams;
+    if (date.isBefore(moment.utc().toDate())) {
+      return next(new InvalidArgumentError(
+        `Invalid date specified: date (${date.toISOString()}) must be in the future.`
+      ));
+    }
+  }
+  next();
+};
+
 const checkPrivilege = function checkPrivilege (req, res, next) {
   if (req.user && req.user.role === config.get('management_role')) {
     return next();
@@ -535,5 +571,6 @@ module.exports = {
   retrieveParameters,
   initUserParams,
   parseAndValidateTimeParamsForFindAllBoxes,
-  checkPrivilege
+  checkPrivilege,
+  validateDateNotPast
 };