+++ packages/models/src/box/box.js
@@ -65,7 +65,7 @@ const boxSchema = new Schema({
     enum: ['unknown', 'indoor', 'outdoor', 'mobile']
   },
   grouptag: {
-    type: String,
+    type: [String], // Default value for array is [] (empty array)
     trim: true,
     required: false
   },
@@ -497,7 +497,9 @@ boxSchema.methods.saveMeasurement = function saveMeasurement (measurement) {
 boxSchema.methods.sensorIds = function sensorIds () {
   const sensorIds = [];
   for (let i = this.sensors.length - 1; i >= 0; i--) {
-    sensorIds.push(this.sensors[i]._id.toString());
+    if (this.sensors[i]._id) {
+      sensorIds.push(this.sensors[i]._id.toString());
+    }
   }
 
   return sensorIds;
@@ -727,7 +729,7 @@ boxSchema.statics.findMeasurementsOfBoxesStream = function findMeasurementsOfBox
       // store all matching sensors under sensors[sensorId]
       for (let i = 0, len = boxData.length; i < len; i++) {
         for (let j = 0, sensorslen = boxData[i].sensors.length; j < sensorslen; j++) {
-          if (boxData[i].sensors[j][sensorProperty].toString() === phenomenon) {
+          if (boxData[i].sensors[j][sensorProperty] && boxData[i].sensors[j][sensorProperty].toString() === phenomenon) {
             const sensor = boxData[i].sensors[j];
 
             sensor.lat = boxData[i].currentLocation.coordinates[1];
@@ -766,6 +768,46 @@ boxSchema.statics.findMeasurementsOfBoxesStream = function findMeasurementsOfBox
         .map(transformer);
     });
 };
+
+boxSchema.statics.findMeasurementsOfBoxesByTagStream = function findMeasurementsOfBoxesByTagStream (opts) {
+  const { query } = opts;
+
+  return this.find(query, BOX_PROPS_FOR_POPULATION)
+    .lean()
+    .then(function (boxData) {
+      if (boxData.length === 0) {
+        throw new ModelError('No senseBoxes found', { type: 'NotFoundError' });
+      }
+
+      const sensors = Object.create(null);
+      // store all matching sensors under sensors[sensorId]
+      for (let i = 0, len = boxData.length; i < len; i++) {
+        for (let j = 0, sensorslen = boxData[i].sensors.length; j < sensorslen; j++) {
+          const sensor = boxData[i].sensors[j];
+
+          sensor.boxId = boxData[i]._id.toString();
+          sensor.sensorId = sensor._id.toString();
+
+          sensors[boxData[i].sensors[j]['_id']] = sensor;
+        }
+      }
+
+      // // construct a stream transformer applied to queried measurements
+      // // that augments each measure with queried columns (location, ...)
+      // // and applies transformations to timestamps
+      const transformer = measurementTransformer(['sensorId', 'boxId'], sensors, undefined);
+
+      const measureQuery = {
+        'sensor_id': { '$in': Object.keys(sensors) },
+        // 'createdAt': { '$gt': from, '$lt': to }
+      };
+
+      return Measurement.find(measureQuery, { 'createdAt': 1, 'value': 1, 'location': 1, '_id': 0, 'sensor_id': 1 })
+        .cursor({ lean: true, order: 1 })
+        .map(transformer);
+    });
+};
+
 
 // try to add sensors defined in addons to the box. If the sensors already exist,
 // nothing is done.
@@ -880,7 +922,7 @@ boxSchema.methods.updateBox = function updateBox (args) {
   // only grouptag, description and weblink can removed through setting them to empty string ('')
   for (const prop of ['name', 'exposure', 'grouptag', 'description', 'weblink', 'image', 'integrations.mqtt', 'integrations.ttn', 'model', 'useAuth']) {
     if (typeof args[prop] !== 'undefined') {
-      box.set(prop, (args[prop] === '' ? undefined : args[prop]));
+      box.set(prop, ((args[prop] === '' || (Array.isArray(args[prop]) && args[prop].length === 0)) ? undefined : args[prop]));
     }
   }
 
@@ -938,14 +980,18 @@ boxSchema.methods.getLocations = function getLocations ({ format, fromDate, toDa
 };
 
 const buildFindBoxesQuery = function buildFindBoxesQuery (opts = {}) {
-  const { phenomenon, fromDate, toDate, bbox, near, maxDistance } = opts,
+  const { phenomenon, fromDate, toDate, bbox, near, maxDistance, grouptag } = opts,
     query = {};
 
   // simple string parameters
-  for (const param of ['exposure', 'model', 'grouptag']) {
+  for (const param of ['exposure', 'model']) {
     if (opts[param]) {
       query[param] = { '$in': opts[param] };
     }
+  }
+
+  if (grouptag) {
+    query['grouptag'] = { '$all': grouptag };
   }
 
   // bbox search parameter
@@ -959,7 +1005,7 @@ const buildFindBoxesQuery = function buildFindBoxesQuery (opts = {}) {
       '$near': {
         '$geometry': {
           type: 'Point',
-          coordinates: [near.split(',')[0], near.split(',')[1]]
+          coordinates: [near[0], near[1]]
         },
         '$maxDistance': maxDistance ? maxDistance : 1000,
       }
@@ -1042,7 +1088,7 @@ boxSchema.statics.findBoxesLastMeasurements = function findBoxesLastMeasurements
 
           for (let i = 0; i < measurementsLength; i++) { //iterate measurments
             for (const sensor of box.sensors) {
-              if (sensor._id.equals(measurements[i].sensor_id)) {
+              if (sensor._id && sensor._id.equals(measurements[i].sensor_id)) {
 
                 measurements[i].sensor_id = undefined;
                 sensor.lastMeasurement = measurements[i];