+++ packages/models/CHANGELOG.md
@@ -2,6 +2,45 @@
 
 ## Unreleased
 
+## v2.0.3
+
+- Change handling `near` parameter (#694)
+
+## v2.0.2
+
+- Fix `undefined` errors in findMeasurementsForSensors (#692)
+
+## v2.0.1
+
+- Fix `undefined` errors in luftdatenHandler (#691)
+
+## v2.0.0
+
+- Update dependencies (#602, #666)
+- Added `findLatestMeasurementsForSensorsWithCount` static method to Measurement schema (#588)
+- Added `sharedBoxes` functionality (#605)
+
+## v1.2.0
+
+- Extend `luftdatenHandler` (#578)
+- Add workflow to transfer a device to different account (#525)
+
+## v1.1.1
+
+- Update @sensebox/node-sketch-templater to v1.12.1
+
+## v1.1.0
+
+- Add DPS310 sensor
+- Update migrations
+- Update @sensebox/node-sketch-templater to v1.12.0
+
+## v1.0.0
+
+- Update @sensebox/node-sketch-templater to v1.11.2
+- `grouptag` changed from `string` to `array`
+- Allow mqtts nad wss for mqtt-integration
+
 ## v0.0.32-beta.0
 - Allow mqtts nad wss for mqtt-integration